# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Novel Adaptations API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Novel Adaptations",
      "url": "https://noveladaptations.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://noveladaptations.com",
      "description": "Production (apex; www.noveladaptations.com serves the same API)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "na_session",
        "description": "Session cookie set by POST /api/v1/auth/verify (HttpOnly, 30 days)."
      }
    },
    "parameters": {
      "pageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "1-based page number. Defaults to 1."
      },
      "perPageParam": {
        "name": "per_page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 24
        },
        "description": "Results per page. Clamped to 100. Default varies by endpoint."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Not signed in. `{ \"error\": { \"code\": \"unauthorized\", \"message\": \"\u2026\" } }`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "Sign in to continue."
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Signed in, but not an admin.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "forbidden",
                "message": "Admin access required."
              }
            }
          }
        }
      },
...
      "Unprocessable": {
        "description": "Validation failure \u2014 always 422 in this API, never 400.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "validation_error",
                "message": "Enter a valid email address."
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Per-user (or per-IP) rate budget exhausted.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "rate_limited",
                "message": "Too many requests \u2014 try again later."
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "not_found"
              },
              "message": {
                "type": "string",
                "example": "Adaptation not found."
              }
            }
          }
        }
      },
...
      "FeedbackStatusCounts": {
        "type": "object",
        "required": [
          "new",
          "reviewed",
          "done"
        ],
        "properties": {
          "new": {
            "type": "integer",
            "example": 3
          },
          "reviewed": {
            "type": "integer",
            "example": 1
          },
          "done": {
            "type": "integer",
            "example": 9
          }
        }
      },
...
      "PagedFeedback": {
        "type": "object",
        "required": [
          "data",
          "page",
          "per_page",
          "total"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackRow"
            }
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "per_page": {
            "type": "integer",
            "example": 50
          },
          "total": {
            "type": "integer",
            "example": 13
          }
        }
      },
...
      "FeedbackRow": {
        "type": "object",
        "required": [
          "id",
          "type",
          "subject",
          "body",
          "proof_url",
          "email",
          "user_id",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "feature",
              "adaptation_tip",
              "correction",
              "other"
            ]
          },
          "subject": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "proof_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "reviewed",
              "done"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/feedback": {
      "post": {
        "tags": [
          "Feedback"
        ],
        "summary": "Submit feedback",
        "description": "Public \u2014 no login required; the session user is attached when present.\nRate-limited by IP. Subject 3\u2013120 chars, body 10\u20135000 chars.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "subject",
                  "body"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "feature",
                      "adaptation_tip",
                      "correction",
                      "other"
                    ],
                    "example": "adaptation_tip"
                  },
                  "subject": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 120,
                    "example": "Dune: Part Three is filming"
                  },
                  "body": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 5000,
                    "example": "Villeneuve started principal photography in Budapest\u2026"
                  },
                  "proof_url": {
                    "type": "string",
                    "example": "https://deadline.com/dune-part-3-filming",
                    "description": "Must start with http:// or https://."
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Optional contact email."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "id"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "id": {
                      "type": "integer",
                      "example": 25
                    }
                  }
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/api/v1/admin/feedback": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Feedback triage queue",
        "description": "`type` and `status` filters are optional; values outside the enums are\nignored (treated as no filter), not rejected.\n",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "feature",
                "adaptation_tip",
                "correction",
                "other"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "reviewed",
                "done"
              ]
            }
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/perPageParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback page plus per-status counts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "filters",
                    "feedback",
                    "counts"
                  ],
                  "properties": {
                    "filters": {
                      "type": "object",
                      "required": [
                        "type",
                        "status"
                      ],
                      "properties": {
                        "type": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "example": "adaptation_tip"
                        },
                        "status": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "example": "new"
                        }
                      }
                    },
                    "feedback": {
                      "$ref": "#/components/schemas/PagedFeedback"
                    },
                    "counts": {
                      "$ref": "#/components/schemas/FeedbackStatusCounts"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  }
}